requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
httpx==0.28.1
python-dotenv==1.1.0
aiohttp==3.12.10
async-retrying-ng==0.1.3
requests==2.32.3